requirements.txt
pandas==1.2.4
torchvision==0.9.1
torch==1.8.1
numpy==1.18.5
requests==2.25.1
Pillow==8.3.1
scikit_learn==0.24.2
